docker_image/requirements.txt
Flask == 1.0.2
google-cloud-pubsub == 0.40.0
redis == 3.2.1
SQLAlchemy == 1.2.17
pg8000 == 1.13.1
google-cloud-datastore==1.7.3
#google-python-cloud-debugger

